uarts: 32/64 cleanup
We had lots of casts that caused warnings when compiling on RISCV.
Clean them up.

Reviewed-on: http://review.coreboot.org/7066
Tested-by: build bot (Jenkins)

+++ src/cpu/allwinner/a10/uart_console.c
@@ -12,7 +12,7 @@
 
 #include "memmap.h"
 
-unsigned int uart_platform_base(int idx)
+uintptr_t uart_platform_base(int idx)
 {
 	/* UART blocks are mapped 0x400 bytes apart */
 	if (idx < 8)
